btf: move typeDeque to separate file
Split typeDeque into a separate file to reduce the amount of code in types.go.
